move sorting methods into their own module
